Move MetaOnly to model package
